[All] Allow multi-tenanted manager instaces
Allows the GRPCManagerInterface to be used independently of proxied
out-of-process manager instantiation. This still needs the C++ plugin
system to be truly useful, but has immediate application in specific
testing scenarios, where shared access to (the somewhat simpleton) BAL
is required.

This also adds a `-d` flag  to the Python `openassetio-grpc-server`
package, when invoked via the CLI, that will instantiate the default
manager and expose it via the `shared` handle, to any
GRPCManagerInterface initialized with a suitable `channel`.

Signed-off-by: Tom Cowland <tom@foundry.com>